+++ GoalMogul-iOS/App.js
@@ -39,6 +39,7 @@ export default class App extends React.Component {
       require('./src/asset/utils/default_profile.png'),
       require('./src/asset/utils/defaultUserProfile.png'),
       require('./src/asset/utils/meetSetting.png'),
+      require('./src/asset/utils/back.png'),
       require('./src/asset/footer/navigation/home.png'),
       require('./src/asset/footer/navigation/bell.png'),
       require('./src/asset/footer/navigation/meet.png'),