+++ GoalMogul-iOS/src/reducers/MeetReducers.js
@@ -1,6 +1,5 @@
 import _ from 'lodash';
 import R from 'ramda';
-import set from 'lodash/fp/set';
 import {
   MEET_SELECT_TAB,
   MEET_LOADING,
@@ -14,13 +13,17 @@ import {
   SETTING_BLOCK_BLOCK_REQUEST_DONE
 } from '../actions/types';
 
-const TabNames = ['SUGGESTED', 'REQUESTS', 'FRIENDS', 'CONTACTS'];
+import {
+  PROFILE_FETCH_FRIEND_DONE,
+  PROFILE_FETCH_FRIEND_COUNT_DONE
+} from './Profile';
+
 const limit = 20;
 const filter = {
   friends: {
     sortBy: ['alphabetical', 'lastadd']
   }
-}
+};
 
 const INITIAL_STATE = {
   selectedTab: 'suggested',
@@ -70,7 +73,8 @@ const INITIAL_STATE = {
     refreshing: false,
     hasNextPage: undefined,
     limit,
-    skip: 0
+    skip: 0,
+    count: undefined
   },
   contacts: {
     data: [],
@@ -108,7 +112,8 @@ export default (state = INITIAL_STATE, action) => {
       // _.set(newState, [action.payload.type, 'loading'], true)
       // console.log('new state is: ', newState);
       // return { state: newState };
-      return set([action.payload.type, 'loading'], true, state);
+      let newState = _.cloneDeep(state);
+      return _.set(newState, `${action.payload.type}.loading`, true);
     }
 
     // Loading suggested cards done
@@ -127,13 +132,14 @@ export default (state = INITIAL_STATE, action) => {
       // _.set(newState, [type, 'data'], data)
       // _.set(newState, [type, 'loading'], false)
       // return { newState };
+      let newState = _.cloneDeep(state);
+      newState = _.set(newState, `${type}.loading`, false);
 
-      let newState = set([type, 'loading'], false, state);
       if (skip !== undefined) {
-        newState = set([type, 'skip'], skip, newState);
+        newState = _.set(newState, `${type}.skip`, skip);
       }
-      newState = set([type, 'hasNextPage'], hasNextPage, newState);
-      return set([type, 'data'], data, newState);
+      newState = _.set(newState, `${type}.hasNextPage`, hasNextPage);
+      return _.set(newState, `${type}.data`, data);
     }
 
     /**
@@ -187,14 +193,13 @@ export default (state = INITIAL_STATE, action) => {
     // Handle tab refresh
     case MEET_TAB_REFRESH: {
       const { type } = action.payload;
-
-      let newState = set([type, 'loading'], true, state);
-      return set([type, 'refreshing'], true, newState);
+      let newState = _.cloneDeep(state);
+      newState = _.set(newState, `${type}.loading`, true);
+      return _.set(newState, `${type}.refreshing`, true);
     }
 
     // Handle tab refresh
     case MEET_TAB_REFRESH_DONE: {
-      // TODO: update the data
       const { type, data } = action.payload;
       let newState = _.set({ ...state }, `${type}.loading`, false);
       newState = _.set({ ...newState }, `${type}.refreshing`, false);
@@ -225,6 +230,18 @@ export default (state = INITIAL_STATE, action) => {
     case SETTING_BLOCK_BLOCK_REQUEST_DONE: {
       const newFriends = { ...state.friends };
       newFriends.data = R.filter((a) => a._id !== action.payload)(newFriends.data);
+      return { ...state, friends: newFriends };
+    }
+
+    // User fetch friend list in profile
+    case PROFILE_FETCH_FRIEND_DONE: {
+      return { ...state };
+    }
+
+    // fetch friends count when opening profile
+    case PROFILE_FETCH_FRIEND_COUNT_DONE: {
+      const newFriends = _.cloneDeep(state.friends);
+      newFriends.count = action.payload;
       return { ...state, friends: newFriends };
     }