+++ GoalMogul-iOS/src/redux/modules/search/Search.js
@@ -1,3 +1,6 @@
+import R from 'ramda';
+import _ from 'lodash';
+
 const INITIAL_STATE = {
   selectedTab: 'people',
   navigationState: {
@@ -39,12 +42,18 @@ const INITIAL_STATE = {
   searchContent: ''
 };
 
+// Helper Functions
+const dotPath = R.useWith(R.path, [R.split('.')]);
+const propsDotPath = R.useWith(R.ap, [R.map(dotPath), R.of]);
+
 // Constants for search reducers
 export const SEARCH_CHANGE_FILTER = 'search_change_filter';
 export const SEARCH_REQUEST = 'search_request';
 export const SEARCH_REQUEST_DONE = 'search_request_done';
 export const SEARCH_REFRESH_DONE = 'search_refresh_done';
 export const SEARCH_SWITCH_TAB = 'search_switch_tab';
+export const SEARCH_ON_LOADMORE_DONE = 'search_on_loadmore_done';
+export const SEARCH_CLEAR_STATE = 'search_clear_state';
 
 /*
   TODO:
@@ -62,7 +71,7 @@ export default (state = INITIAL_STATE, action) => {
     // Initiate search request
     case SEARCH_REQUEST: {
       const { searchContent, queryId, type } = action.payload;
-      let newState = { ...state };
+      let newState = _.cloneDeep(state);
       newState[type].loading = true;
       newState.queryId = queryId;
       newState.searchContent = searchContent;
@@ -73,9 +82,9 @@ export default (state = INITIAL_STATE, action) => {
     case SEARCH_REFRESH_DONE:
     case SEARCH_REQUEST_DONE: {
       const { queryId, skip, data, type, hasNextPage } = action.payload;
-      let newState = { ...state };
+      let newState = _.cloneDeep(state);
       if (queryId === state.queryId) {
-        newState[type].data = newState[type].data.concat(data);
+        newState[type].data = data;
         newState[type].loading = false;
         newState[type].skip = skip;
         newState[type].hasNextPage = hasNextPage;
@@ -84,19 +93,17 @@ export default (state = INITIAL_STATE, action) => {
     }
 
     // Search refresh done
-    // case SEARCH_REFRESH_DONE: {
-    //   const { queryId, skip, data, type } = action.payload;
-    //
-    //   if (queryId === state.queryId) {
-    //     return {
-    //       ...state,
-    //       loading: false,
-    //       data,
-    //       skip
-    //     };
-    //   }
-    //   return { ...state };
-    // }
+    case SEARCH_ON_LOADMORE_DONE: {
+      const { queryId, skip, data, type, hasNextPage } = action.payload;
+      let newState = _.cloneDeep(state);
+      if (queryId === state.queryId) {
+        newState[type].data = newState[type].data.concat(data);
+        newState[type].loading = false;
+        newState[type].skip = skip;
+        newState[type].hasNextPage = hasNextPage;
+      }
+      return { ...newState };
+    }
 
     // Search switch tab
     case SEARCH_SWITCH_TAB: {
@@ -108,6 +115,18 @@ export default (state = INITIAL_STATE, action) => {
         selectedTab: newNavigationState.routes[action.payload].key,
         navigationState: newNavigationState
       };
+    }
+
+    case SEARCH_CLEAR_STATE: {
+      const { tab } = action.payload;
+      if (!tab) {
+        // No tab specified, clear all state
+        return INITIAL_STATE;
+      }
+      // Clear state for specific tab
+      const tabInitialState = dotPath(tab, INITIAL_STATE);
+      const newState = _.cloneDeep(state);
+      return _.set(newState, tab, tabInitialState);
     }
 
     default: